Updated gamma filter to accept double

+++ filters/src/main/scala/com/sksamuel/scrimage/filter/GammaFilter.scala
@@ -18,11 +18,11 @@ package com.sksamuel.scrimage.filter
 import com.sksamuel.scrimage.BufferedOpFilter
 
 /** @author Stephen Samuel */
-class GammaFilter(gamma: Float) extends BufferedOpFilter {
+class GammaFilter(gamma: Double) extends BufferedOpFilter {
   val op = new thirdparty.jhlabs.image.GammaFilter()
-  op.setGamma(gamma)
+  op.setGamma(gamma.toFloat)
 }
 object GammaFilter {
   def apply(): GammaFilter = apply(1.0f)
-  def apply(gamma: Float): GammaFilter = new GammaFilter(gamma)
+  def apply(gamma: Double): GammaFilter = new GammaFilter(gamma)
 }
